Refactor: move GetServiceAccountSecretName to new ServiceAccountHelper
