+++ pkg/k8s/serviceAccountManager_test.go
@@ -1,9 +1,7 @@
 package k8s
 
 import (
-	"sync"
 	"testing"
-	"time"
 
 	"github.com/SAP/stewardci-core/pkg/k8s/fake"
 	"gotest.tools/assert"
@@ -108,153 +106,4 @@ func Test_SetDoAutomountServiceAccountToken_works(t *testing.T) {
 	actual, err := accountManager.GetServiceAccount(accountName)
 	assert.NilError(t, err)
 	assert.Check(t, *actual.cache.AutomountServiceAccountToken == false)
-}
-
-func Test_GetServiceAccountSecretName_works(t *testing.T) {
-	//SETUP
-	secretName := "ns1-token-foo"
-	secret := &v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      secretName,
-			Namespace: ns1,
-		},
-		Type: v1.SecretTypeServiceAccountToken,
-	}
-	setupAccountManager(secret)
-	acc, err := accountManager.CreateServiceAccount(accountName, "pipelineCloneSecretName1", []string{"imagePullSecret1", "imagePullSecret2"})
-	assert.NilError(t, err)
-
-	acc.AttachSecrets("a-secret", secretName, "z-secret")
-
-	// EXERCISE
-	resultName := acc.GetServiceAccountSecretName()
-	// VERIFY
-	assert.Equal(t, secretName, resultName)
-}
-
-func Test_GetServiceAccountSecretNameRepeat_delayedRef_works(t *testing.T) {
-	//SETUP
-	secretName := "ns1-token-foo"
-	secret := &v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      secretName,
-			Namespace: ns1,
-		},
-		Type: v1.SecretTypeServiceAccountToken,
-	}
-	setupAccountManager(secret)
-	acc, err := accountManager.CreateServiceAccount(accountName, "pipelineCloneSecretName1", []string{"imagePullSecret1", "imagePullSecret2"})
-	assert.NilError(t, err)
-	err = acc.Update()
-	assert.NilError(t, err)
-	var waitGroup sync.WaitGroup
-	waitGroup.Add(1)
-	go func(t *testing.T, acc *ServiceAccountWrap) {
-		defer waitGroup.Done()
-		// EXERCISE
-		resultName := acc.GetServiceAccountSecretNameRepeat()
-		// VERIFY
-		assert.Equal(t, "ns1-token-foo", resultName)
-	}(t, acc)
-	duration, _ := time.ParseDuration("500ms")
-	time.Sleep(duration)
-	localAccountManager := NewServiceAccountManager(factory, ns1)
-	localAccount, err := localAccountManager.GetServiceAccount(accountName)
-	assert.NilError(t, err)
-	localAccount.AttachSecrets("a-secret", secretName, "z-secret")
-	err = localAccount.Update()
-	assert.NilError(t, err)
-	waitGroup.Wait()
-}
-
-func Test_GetServiceAccountSecretNameRepeat_delayedSecret_works(t *testing.T) {
-	//SETUP
-	secretName := "ns1-token-foo"
-	secret := &v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      secretName,
-			Namespace: ns1,
-		},
-		Type: v1.SecretTypeServiceAccountToken,
-	}
-	setupAccountManager()
-	acc, err := accountManager.CreateServiceAccount(accountName, "pipelineCloneSecretName1", []string{"imagePullSecret1", "imagePullSecret2"})
-	acc.AttachSecrets("a-secret", secretName, "z-secret")
-	assert.NilError(t, err)
-	err = acc.Update()
-	assert.NilError(t, err)
-	var waitGroup sync.WaitGroup
-	waitGroup.Add(1)
-	go func(t *testing.T, acc *ServiceAccountWrap) {
-		defer waitGroup.Done()
-		// EXERCISE
-		resultName := acc.GetServiceAccountSecretNameRepeat()
-		// VERIFY
-		assert.Equal(t, "ns1-token-foo", resultName)
-	}(t, acc)
-	duration, _ := time.ParseDuration("500ms")
-	time.Sleep(duration)
-	secretsInterface := factory.CoreV1().Secrets(ns1)
-	_, err = secretsInterface.Create(secret)
-	assert.NilError(t, err)
-	waitGroup.Wait()
-}
-
-func Test_GetServiceAccountSecretName_wrongType(t *testing.T) {
-	//SETUP
-	secretName := "ns1-token-foo"
-	secret := &v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      secretName,
-			Namespace: ns1,
-		},
-		Type: v1.SecretTypeOpaque,
-	}
-	setupAccountManager(secret)
-	acc, err := accountManager.CreateServiceAccount(accountName, "pipelineCloneSecretName1", []string{"imagePullSecret1", "imagePullSecret2"})
-	assert.NilError(t, err)
-
-	acc.AttachSecrets("a-secret", secretName, "z-secret")
-
-	// EXERCISE
-	resultName := acc.GetServiceAccountSecretName()
-	// VERIFY
-	assert.Equal(t, "", resultName)
-}
-
-func Test_GetServiceAccountSecretName_ref_missing(t *testing.T) {
-	//SETUP
-	secretName := "ns1-token-foo"
-	secret := &v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      secretName,
-			Namespace: ns1,
-		},
-		Type: v1.SecretTypeServiceAccountToken,
-	}
-	setupAccountManager(secret)
-	acc, err := accountManager.CreateServiceAccount(accountName, "pipelineCloneSecretName1", []string{"imagePullSecret1", "imagePullSecret2"})
-	assert.NilError(t, err)
-
-	acc.AttachSecrets("a-secret", "z-secret")
-
-	// EXERCISE
-	resultName := acc.GetServiceAccountSecretName()
-	// VERIFY
-	assert.Equal(t, "", resultName)
-}
-
-func Test_GetServiceAccountSecretName_secret_missing(t *testing.T) {
-	//SETUP
-	secretName := "ns1-token-foo"
-	setupAccountManager()
-	acc, err := accountManager.CreateServiceAccount(accountName, "pipelineCloneSecretName1", []string{"imagePullSecret1", "imagePullSecret2"})
-	assert.NilError(t, err)
-
-	acc.AttachSecrets("a-secret", secretName, "z-secret")
-
-	// EXERCISE
-	resultName := acc.GetServiceAccountSecretName()
-	// VERIFY
-	assert.Equal(t, "", resultName)
 }
